+++ packages/web/src/Components/Popover.tsx
@@ -1,5 +1,5 @@
 import React, { FunctionComponent, forwardRef } from 'react';
-import { withTheme } from '@material-ui/core';
+import { PopperPlacementType, withTheme } from '@material-ui/core';
 import MaterialPopper, {
   PopperProps as IPopperProps,
 } from '@material-ui/core/Popper';
@@ -7,11 +7,39 @@ import MaterialPopper, {
 import styled from 'styled-components';
 import { IThemeWeb } from 'Themes';
 import { getDefaultTheme } from './shared';
+import { Paper, Typography, Link, ILinkProps } from '../';
 
-export interface IPopoverProps extends IPopperProps {
+interface IActionLink extends Omit<ILinkProps, 'theme'> {
+  text: string;
+  href?: string;
+  onClick?: any;
+}
+
+export interface IPopoverProps extends Omit<IPopperProps, 'placement'> {
   theme: IThemeWeb | unknown;
   component?: React.ElementType;
+  actionLink?: IActionLink;
+  direction?: PopperPlacementType;
+  maxWidth?: number;
 }
+
+const getActionLink = (actionLink?: IActionLink) => {
+  if (actionLink) {
+    const { text, ...props } = actionLink;
+    return (
+      actionLink && (
+        <Link
+          {...props}
+          variant="body2"
+          className="popover-action-link"
+        >
+          {text}
+        </Link>
+      )
+    );
+  }
+  return null;
+};
 
 export const Popover: FunctionComponent<IPopoverProps> = forwardRef(
   (props: IPopoverProps, ref: any) => {
@@ -20,6 +48,8 @@ export const Popover: FunctionComponent<IPopoverProps> = forwardRef(
       component,
       theme: providerTheme,
       modifiers,
+      actionLink,
+      direction = 'bottom',
       ...rest
     } = props;
     const theme: any = React.useMemo(() => getDefaultTheme(providerTheme), [
@@ -47,10 +77,14 @@ export const Popover: FunctionComponent<IPopoverProps> = forwardRef(
             boundariesElement: 'scrollParent',
           },
         }}
+        placement={direction}
         {...rest}
       >
         <span className="arrow" ref={setArrowRef} />
-        {children}
+        <Paper elevation={12} variant="elevation" className="popover-container">
+          <Typography variant="body2">{children}</Typography>
+          {getActionLink(actionLink)}
+        </Paper>
       </StyledPopper>
     );
   }
@@ -58,60 +92,72 @@ export const Popover: FunctionComponent<IPopoverProps> = forwardRef(
 
 export default withTheme(Popover);
 
-const StyledPopper = styled(MaterialPopper)<{ theme: IThemeWeb }>`
+const StyledPopper = styled(MaterialPopper)<{ theme: IThemeWeb, maxWidth: number }>`
   && {
     z-index: 1;
-    &[x-placement*='bottom'] .arrow {
-      width: 3em;
-      height: 1em;
-      margin-top: -0.9em;
+    &[x-placement*='bottom'] {
+      margin-top: ${({ theme }) => theme.spacing(1)}px;
+      & .arrow {
+        width: 3em;
+        height: 1em;
+        margin-top: -0.9em;
 
-      &:before {
-        border-width: 0 1em 1em 1em;
-        border-color: ${({ theme }) =>
-          `transparent transparent ${theme.palette.background.paper} transparent`};
+        &:before {
+          margin-left: 3px;
+          border-width: 0 1em 1em 1em;
+          border-color: ${({ theme }) =>
+            `transparent transparent ${theme.palette.background.paper} transparent`};
+        }
       }
     }
+    &[x-placement*='top'] {
+      margin-bottom: ${({ theme }) => theme.spacing(1)}px;
+      & .arrow {
+        bottom: 0;
+        width: 3em;
+        height: 1em;
+        margin-bottom: -0.9em;
 
-    &[x-placement*='top'] .arrow {
-      bottom: 0;
-      width: 3em;
-      height: 1em;
-      margin-bottom: -0.9em;
-
-      &:before {
-        border-width: 1em 1em 0 1em;
-        border-color: ${({ theme }) =>
-          `${theme.palette.background.paper} transparent transparent transparent`};
+        &:before {
+          margin-left: 3px;
+          border-width: 1em 1em 0 1em;
+          border-color: ${({ theme }) =>
+            `${theme.palette.background.paper} transparent transparent transparent`};
+        }
       }
     }
+    &[x-placement*='right'] {
+      margin-left: ${({ theme }) => theme.spacing(1)}px;
+      & .arrow {
+        left: 0;
+        width: 1em;
+        height: 3em;
+        margin-left: -0.9em;
 
-    &[x-placement*='right'] .arrow {
-      left: 0;
-      width: 1em;
-      height: 3em;
-      margin-left: -0.9em;
-
-      &:before {
-        border-width: 1em 1em 1em 0;
-        border-color: ${({ theme }) =>
-          `transparent ${theme.palette.background.paper} transparent transparent`};
+        &:before {
+          margin-top: 3px;
+          border-width: 1em 1em 1em 0;
+          border-color: ${({ theme }) =>
+            `transparent ${theme.palette.background.paper} transparent transparent`};
+        }
       }
     }
+    &[x-placement*='left'] {
+      margin-right: ${({ theme }) => theme.spacing(1)}px;
+      & .arrow {
+        right: 0;
+        width: 1em;
+        height: 3em;
+        margin-right: -0.9em;
 
-    &[x-placement*='left'] .arrow {
-      right: 0;
-      width: 1em;
-      height: 3em;
-      margin-right: -0.9em;
-
-      &:before {
-        border-width: 1em 0 1em 1em;
-        border-color: ${({ theme }) =>
-          `transparent transparent transparent ${theme.palette.background.paper}`};
+        &:before {
+          margin-top: 3px;
+          border-width: 1em 0 1em 1em;
+          border-color: ${({ theme }) =>
+            `transparent transparent transparent ${theme.palette.background.paper}`};
+        }
       }
     }
-
     .arrow {
       position: absolute;
       font-size: 7px;
@@ -126,6 +172,16 @@ const StyledPopper = styled(MaterialPopper)<{ theme: IThemeWeb }>`
         height: 0;
         border-style: solid;
       }
+    }
+
+    .popover-container {
+      max-width: ${({ maxWidth }) => maxWidth || 380}px;
+      padding: ${({ theme }) => theme.spacing(2)}px;
+    }
+
+    .popover-action-link {
+      display: inline-block;
+      padding-top: ${({ theme }) => theme.spacing(1)}px;
     }
   }
 `;