+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,58 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.1](https://github.com/natura-cosmeticos/natds/compare/@naturacosmeticos/natds-icons@0.5.0...@naturacosmeticos/natds-icons@0.5.1) (2020-04-30)
+
+**Note:** Version bump only for package @naturacosmeticos/natds-icons
+
+
+
+
+
+# [0.5.0](https://github.com/natura-cosmeticos/natds/compare/@naturacosmeticos/natds-icons@0.4.5...@naturacosmeticos/natds-icons@0.5.0) (2020-04-29)
+
+
+### Features
+
+* **Component Breadcrumbs:** creating new component Breadcrombs ([772d711](https://github.com/natura-cosmeticos/natds/commit/772d7112469899eb6b83b31994d3c878418c7f20))
+
+
+
+
+
+## [0.4.5](https://github.com/natura-cosmeticos/natds/compare/@naturacosmeticos/natds-icons@0.4.4...@naturacosmeticos/natds-icons@0.4.5) (2020-04-28)
+
+
+### Bug Fixes
+
+* **outlined-action-icons:** included new outlined action icons ([1f482ac](https://github.com/natura-cosmeticos/natds/commit/1f482acea721b95d033aa17a4ca9a7bc1c61104e))
+
+
+
+
+
+## [0.4.4](https://github.com/natura-cosmeticos/natds/compare/@naturacosmeticos/natds-icons@0.4.3...@naturacosmeticos/natds-icons@0.4.4) (2020-04-23)
+
+
+### Bug Fixes
+
+* **icons:** added new filled and outlined icons ([7055041](https://github.com/natura-cosmeticos/natds/commit/70550417f90416e3384883ca44d5214330dd2926))
+
+
+
+
+
+## [0.4.3](https://github.com/natura-cosmeticos/natds/compare/@naturacosmeticos/natds-icons@0.4.2...@naturacosmeticos/natds-icons@0.4.3) (2020-04-15)
+
+
+### Bug Fixes
+
+* **icon:** added new icon named outlined-content-indicator ([fa8670d](https://github.com/natura-cosmeticos/natds/commit/fa8670d6d703e6adccacbfad7dacaae78e0b06b8))
+
+
+
+
+
 ## [0.4.2](https://github.com/natura-cosmeticos/natds/compare/@naturacosmeticos/natds-icons@0.4.1...@naturacosmeticos/natds-icons@0.4.2) (2020-04-13)